chore: Create common context object
